move hook functions + layers_info fn to utils file
