+++ .gitignore
@@ -1,3 +1,6 @@
 *.csv
 venv/
 *.txt
+__pycache__/
+.DS_Store
+.ipynb_checkpoints/
