+++ khoa/algorithms/iterative_relaxation.py
@@ -1,9 +1,9 @@
 
 
-from algorithms.decision_procedure import MarabouCoreDP
 import copy
 import torch
-from torch import nn
+from algorithms.decision_procedure import MarabouCoreDP
+from models.utils import attach_relu_activation_hook, attach_layer_output_hook
 
 class IterativeRelaxation():
   def __init__(self):
@@ -11,8 +11,8 @@ class IterativeRelaxation():
 
   def call(self, model, input_data, postcondition):
     # attach hooks to model to get activation signature of X
-    _act_handles, activation_signature = model.attach_relu_activation_hook()
-    _out_handles, layer_outputs = model.attach_layer_output_hook()
+    _act_handles, activation_signature = attach_relu_activation_hook(model)
+    _out_handles, layer_outputs = attach_layer_output_hook(model)
 
     # evaluate model with X to get activation signature of X
     X = torch.tensor(input_data, dtype=torch.float)
